requirements.txt
Flask-Mail==0.9.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
